+++ R/functions-msp.R
@@ -1,43 +1,54 @@
-##' @param f `character(1)` with the path to an msp file.
-##' 
-##' @param msLevel `numeric(1)` with the MS level. Default is 2.
-##' 
-##' @param ... Additional parameters, currently ignored.
-##'
-##' @importFrom S4Vectors DataFrame
-##'
-##' @importFrom IRanges NumericList
-##' 
-##' @author Laurent Gatto
-##' 
-##' @noRd
-
-if (FALSE) {
-    library(MSnbase)
-    library(microbenchmark)
-    f <- "/vol/bioinvindex/Submissions/MIC044/MTBLS1582/20203111526_spectra_SMpos.msp"
-    f <- "/vol/R/BioC/devel/Spectra/MassBank_MSMS_Pos_Rev173_vs1.msp"
-    f <- "/vol/R/BioC/devel/2018-03-09_14_10_01_pos_387850_nist_msms_HR.MSP"
-    f <- "/vol/R/BioC/devel/MsBackendMsp/inst/extdata/Spectrum.msp"
-    f <- "/vol/R/BioC/devel/MsBackendMsp/inst/extdata/Spectrum2.msp"
-    f <- "/vol/R/BioC/devel/MsBackendMsp/inst/extdata/msdial_pos.msp"
-    f <- "/vol/R/BioC/devel/MsBackendMsp/inst/extdata/first-export-LipidBlast.msp"
-    massbank <- .read_msp(f)
-    orgmassbank <- ReadMspFile(f)
-    
-    microbenchmark(.read_msp(f), times = 5)
-}
-
-
-.read_msp <- function(f, msLevel = 2L, ...) {
+#' @title Reading MSP files
+#'
+#' The `readMsp` function imports the data from a file in MGF format reading
+#' all specified fields and returning the data as a [DataFrame()].
+#'
+#' @param f `character(1)` with the path to an MSP file.
+#'
+#' @param msLevel `numeric(1)` with the MS level. Default is 2. This value will
+#'     be reported as the spectra's MS level **unless** the source MSP file
+#'     defines the MS level.
+#'
+#' @param mapping named `character` vector to rename MSP fields to spectra
+#'     variables (see [spectraVariableMapping()]). This allows to correctly
+#'     import also custom fields or data from files with different MSP
+#'     *flavors*.
+#'
+#' @param ... Additional parameters, currently ignored.
+#'
+#' @return
+#'
+#' A `DataFrame` with each row containing the data from one spectrum
+#' in the MSP file. m/z and intensity values are available in columns `"mz"`
+#' and `"intensity"` in a list representation.
+#'
+#' @export
+#'
+#' @importFrom S4Vectors DataFrame
+#'
+#' @importFrom IRanges NumericList
+#'
+#' @importFrom MsCoreUtils rbindFill
+#'
+#' @importFrom methods as
+#'
+#' @author Laurent Gatto, Steffen Neumann, Johannes Rainer
+#'
+#' @examples
+#'
+#' fls <- dir(system.file("extdata", package = "MsBackendMsp"),
+#'     full.names = TRUE, pattern = "msp$")[1L]
+#'
+#' readMsp(fls)
+readMsp <- function(f, msLevel = 2L,
+                    mapping = spectraVariableMapping(MsBackendMsp()), ...) {
     if (length(f) != 1L)
         stop("Please provide a single msp file.")
     
     msp <- scan(file = f, what = "",
                 sep = "\n", quote = "",
                 allowEscapes = FALSE,
-                quiet = TRUE
-                )
+                quiet = TRUE)
     
     ## Ignore comments
     cmts <- grep("^[#]", msp)
@@ -46,82 +57,93 @@ if (FALSE) {
 
     ## Find individual records
     begin <- grep("^NAME:", msp, ignore.case = TRUE)
-    stopifnot(!isEmpty(begin)) ## NO! DONT STOP ON ERROR! FIND OUT HOW EXCEPTIONS WORK!
-    
-    end <- c(begin[-1], length(msp))
+    end <- c(begin[-1] -1L, length(msp))
     
     n <- length(begin)
     sp <- vector("list", length = n)
 
-    for (i in seq(along = sp)) 
-        sp[[i]] <- .extract_msp_spectrum(msp[begin[i]:end[i]])
+    for (i in seq_along(sp))
+        sp[[i]] <- .extract_msp_spectrum(msp[begin[i]:end[i]],
+                                         mapping = mapping)
+    res <- DataFrame(rbindFill(sp))
 
-    res <- DataFrame(do.call(rbind, sp))
-
+    spv <- Spectra:::.SPECTRA_DATA_COLUMNS
+    spv <- spv[!names(spv) %in% c("mz", "intensity")]
     for (i in seq_along(res)) {
         if (all(lengths(res[[i]]) == 1))
             res[[i]] <- unlist(res[[i]])
+        if (any(col <- names(spv) == colnames(res)[i]))
+            res[[i]] <- as(res[[i]], spv[col][1])
     }
 
-    res$mz <- IRanges::NumericList(res$mz)
-    res$intensity <- IRanges::NumericList(res$intensity)
+    res$mz <- IRanges::NumericList(res$mz, compress = FALSE)
+    res$intensity <- IRanges::NumericList(res$intensity, compress = FALSE)
     res$dataOrigin <- f
-    res$msLevel <- as.integer(msLevel)
+    if (!any(colnames(res) == msLevel))
+        res$msLevel <- as.integer(msLevel)
     res
 }
 
-##' @param msp `character()` of lines defining a spectrum in msp
-##'     format.
-##' 
-##' @author Laurent Gatto
-##' 
-##' @importFrom stats setNames
-##'
-##' @noRd
-.extract_msp_spectrum <- function(msp) {
+#' @param msp `character()` of lines defining a spectrum in msp
+#'     format.
+#'
+#' @param mapping spectra variable mapping that allows renaming data fields.
+#' 
+#' @author Laurent Gatto, Johannes Rainer
+#' 
+#' @importFrom stats setNames
+#'
+#' @noRd
+.extract_msp_spectrum <- function(msp, mapping) {
     ## grep description
     desc.idx <- grep(":", msp)
     desc <- msp[desc.idx]
     spec <- msp[-desc.idx]
 
-    ms <- do.call(rbind, strsplit(spec, "[[:space:]]+"))
+    ms <- do.call(rbind, strsplit(sub("^(\\t|[[:space:]]+)", "", spec),
+                                  "[[:space:]]+"))
     mode(ms) <- "double"
 
     if (!length(ms))
         ms <- matrix(numeric(), ncol = 2L)
-    
-    if(length(ms) > 2)
-        ms <- ms[order(ms[, 1L]),]
+    else if (is.unsorted(ms[, 1L]))
+        ms <- ms[order(ms[, 1L]), ]
     
     r <- regexpr(":", desc, fixed = TRUE)
-    desc <- setNames(substring(desc, r + 2L, nchar(desc)), tolower(substring(desc, 1L, r - 1L)))
-    name <- unname(desc["name"])
-    accession <- unname(desc["db#"])
-    formula <- unname(desc["formula"])
-    inchikey <- unname(desc["inchikey"])
-    adduct <- unname(desc["precursor_type"])
-    exactmass <- as.numeric(unname(desc["exactmass"]))
-    comments <- unname(desc["comments"])
-    ion_mode <- unname(desc["ion_mode"])
+    desc <- setNames(substring(desc, r + 2L, nchar(desc)),
+                     substring(desc, 1L, r - 1L))
+    ## map fields to spectra variables
+    idx <- match(names(desc), mapping)
+    not_na <- !is.na(idx)
+    if (any(not_na))
+        names(desc)[not_na] <- names(mapping)[idx][not_na]
+
+    ## Handle eventually duplicated names -> list
+    if (anyDuplicated(names(desc))) {
+        res <- split(unname(desc), names(desc))
+        dups <- lengths(res) > 1L
+        dup_res <- res[dups]
+        res <- as.data.frame(res[!dups])
+        for (name in names(dup_res))
+            res <- do.call("$<-", list(res, name, unname(dup_res[name])))
+    } else res <- as.data.frame(as.list(desc))
     
-    ## select only values of interest and convert to numeric
-    voi <- c("retentiontime", "ionmode", "precursormz")
-    desc <- setNames(as.numeric(desc[voi]), voi)
-    desc[is.na(desc[voi])] <- 0L
-    
-    list(rtime = unname(desc["retentiontime"]),
-         precursorMz = unname(desc["precursormz"]),
-         precursorCharge = unname(as.integer(desc["ionmode"])),
-         mz = ms[, 1L],
-         intensity = ms[, 2L],
-         accession = accession,
-         name = name,
-         formula = formula,
-         inchikey = inchikey,
-         adduct = adduct,
-         exactmass = exactmass,
-         comments = comments,
-         ion_mode = ion_mode)
+    ## Ensure correct data type
+    ## polarity
+    if (any(have <- colnames(res) == "polarity"))
+        res[, have] <- .process_polarity(res[, have])
+
+    res$mz = list(ms[, 1L])
+    res$intensity = list(ms[, 2L])
+    res
+}
+
+.process_polarity <- function(x) {
+    if (grepl("^(p|\\+)", x, ignore.case = TRUE))
+        return(1L)
+    if (grepl("^(n|-)", x, ignore.case = TRUE))
+        return(0L)
+    -1L
 }
 
 #' @description
@@ -138,7 +160,9 @@ if (FALSE) {
 #'
 #' @author Michael Witting
 #'
-#' @importMethodsFrom Spectra spectraVariables spectraNames spectraData
+#' @importMethodsFrom Spectra spectraVariables spectraNames
+#'
+#' @importMethodsFrom Spectra spectraData peaksData
 #'
 #' @noRd
 .export_msp <- function(x, con = stdout(), mapping = spectraVariableMapping()) {