+++ NEWS.md
@@ -1,3 +1,11 @@
+# MsBackendMsp 0.0
+
+## Changes in 0.0.5
+
+- Add possibility to define mapping between spectra variables and MSP data 
+  fields.
+- Support duplicated data fields in a single MSP entry.
+
 ## Changes in 0.0.4
 
 - Fix tests to match new columns in MSP parsing